Extract web fetch example into module
